[CHANGE] views split up into multiple files
