+++ aasrp/urls.py
@@ -6,49 +6,52 @@ aasrp url config
 from django.urls import include, path
 
 # AA SRP
-from aasrp import views
+from aasrp.views import ajax, general
 
 app_name: str = "aasrp"
 
 urlpatterns = [
-    path(route="", view=views.dashboard, name="dashboard"),
+    path(route="", view=general.dashboard, name="dashboard"),
     path(
-        route="all/", view=views.dashboard, kwargs={"show_all_links": True}, name="all"
+        route="all/",
+        view=general.dashboard,
+        kwargs={"show_all_links": True},
+        name="all",
     ),
-    path(route="add/", view=views.srp_link_add, name="add_srp_link"),
+    path(route="add/", view=general.srp_link_add, name="add_srp_link"),
     path(
         route="srp-link/<str:srp_code>/edit/",
-        view=views.srp_link_edit,
+        view=general.srp_link_edit,
         name="edit_srp_link",
     ),
     path(
         route="srp-link/<str:srp_code>/view-srp-requests/",
-        view=views.srp_link_view_requests,
+        view=general.srp_link_view_requests,
         name="view_srp_requests",
     ),
     path(
         route="srp-link/<str:srp_code>/enable/",
-        view=views.enable_srp_link,
+        view=general.enable_srp_link,
         name="enable_srp_link",
     ),
     path(
         route="srp-link/<str:srp_code>/disable/",
-        view=views.disable_srp_link,
+        view=general.disable_srp_link,
         name="disable_srp_link",
     ),
     path(
         route="srp-link/<str:srp_code>/delete/",
-        view=views.delete_srp_link,
+        view=general.delete_srp_link,
         name="delete_srp_link",
     ),
     path(
         route="srp-link/<str:srp_code>/request-srp/",
-        view=views.request_srp,
+        view=general.request_srp,
         name="request_srp",
     ),
     path(
         route="srp-link/<str:srp_code>/complete/",
-        view=views.complete_srp_link,
+        view=general.complete_srp_link,
         name="complete_srp_link",
     ),
     # Ajax calls
@@ -59,26 +62,26 @@ urlpatterns = [
                 path(
                     # Get active SRP links
                     route="active-srp-links-data/",
-                    view=views.ajax_dashboard_srp_links_data,
+                    view=ajax.dashboard_srp_links_data,
                     name="ajax_dashboard_srp_links_data",
                 ),
                 path(
                     # Get all SRP links
                     route="active-srp-links-data/all/",
-                    view=views.ajax_dashboard_srp_links_data,
+                    view=ajax.dashboard_srp_links_data,
                     kwargs={"show_all_links": True},
                     name="ajax_dashboard_srp_links_all_data",
                 ),
                 path(
                     # Get all SRP request for the current user
                     route="user-srp-requests-data/",
-                    view=views.ajax_dashboard_user_srp_requests_data,
+                    view=ajax.dashboard_user_srp_requests_data,
                     name="ajax_dashboard_user_srp_requests_data",
                 ),
                 path(
                     # Get all SRP requests for the current SRP link
                     route="srp-link/<str:srp_code>/view-srp-requests-data/",
-                    view=views.ajax_srp_link_view_requests_data,
+                    view=ajax.srp_link_view_requests_data,
                     name="ajax_srp_link_view_requests_data",
                 ),
                 path(
@@ -87,7 +90,7 @@ urlpatterns = [
                         "srp-link/<str:srp_code>/srp-request/"
                         "<str:srp_request_code>/view-additional-information-data/"
                     ),
-                    view=views.ajax_srp_request_additional_information,
+                    view=ajax.srp_request_additional_information,
                     name="ajax_srp_request_additional_information",
                 ),
                 path(
@@ -96,7 +99,7 @@ urlpatterns = [
                         "srp-link/<str:srp_code>/srp-request/"
                         "<str:srp_request_code>/change-srp-payout/"
                     ),
-                    view=views.ajax_srp_request_change_payout,
+                    view=ajax.srp_request_change_payout,
                     name="ajax_srp_request_change_payout",
                 ),
                 path(
@@ -105,7 +108,7 @@ urlpatterns = [
                         "srp-link/<str:srp_code>/srp-request/"
                         "<str:srp_request_code>/approve/"
                     ),
-                    view=views.ajax_srp_request_approve,
+                    view=ajax.srp_request_approve,
                     name="ajax_srp_request_approve",
                 ),
                 path(
@@ -114,7 +117,7 @@ urlpatterns = [
                         "srp-link/<str:srp_code>/srp-request/"
                         "<str:srp_request_code>/deny/"
                     ),
-                    view=views.ajax_srp_request_deny,
+                    view=ajax.srp_request_deny,
                     name="ajax_srp_request_deny",
                 ),
                 path(
@@ -123,7 +126,7 @@ urlpatterns = [
                         "srp-link/<str:srp_code>/srp-request/"
                         "<str:srp_request_code>/remove/"
                     ),
-                    view=views.ajax_srp_request_remove,
+                    view=ajax.srp_request_remove,
                     name="ajax_srp_request_remove",
                 ),
             ]
